refactor(queues): move negotiation-claim-timeout into negotiations/ subdir
