+++ backend/src/main.ts
@@ -44,7 +44,7 @@ import { hydeQueue } from './queues/hyde.queue';
 import { emailQueue } from './queues/email.queue';
 import { profileQueue } from './queues/profile.queue';
 import { negotiationTimeoutQueue } from './queues/negotiation-timeout.queue';
-import { negotiationClaimTimeoutQueue } from './queues/negotiation-claim-timeout.queue';
+import { negotiationClaimTimeoutQueue } from './queues/negotiations/claim-timeout.queue';
 import { NetworkMembershipEvents } from './events/network_membership.event';
 import { IntentEvents } from './events/intent.event';
 import { NegotiationEvents } from './events/negotiation.event';
