+++ backend/src/queues/negotiations/claim-timeout.queue.ts
@@ -1,11 +1,11 @@
 import { Job } from 'bullmq';
 import { and, eq } from 'drizzle-orm';
 
-import db from '../lib/drizzle/drizzle';
-import * as convSchema from '../schemas/conversation.schema';
-import { log } from '../lib/log';
-import { QueueFactory } from '../lib/bullmq/bullmq';
-import { conversationDatabaseAdapter } from '../adapters/database.adapter';
+import db from '../../lib/drizzle/drizzle';
+import * as convSchema from '../../schemas/conversation.schema';
+import { log } from '../../lib/log';
+import { QueueFactory } from '../../lib/bullmq/bullmq';
+import { conversationDatabaseAdapter } from '../../adapters/database.adapter';
 import { IndexNegotiator, AMBIENT_PARK_WINDOW_MS } from '@indexnetwork/protocol';
 import type { NegotiationTurn, NegotiationOutcome, UserNegotiationContext, SeedAssessment, NegotiationDatabase } from '@indexnetwork/protocol';
 
@@ -313,7 +313,7 @@ export class NegotiationClaimTimeoutQueue {
     await database.updateTaskState(task.id, 'waiting_for_agent');
 
     // Import dynamically to avoid circular dependency; arm the park-window timeout for the next speaker
-    const { negotiationTimeoutQueue } = await import('./negotiation-timeout.queue');
+    const { negotiationTimeoutQueue } = await import('../negotiation-timeout.queue');
     await negotiationTimeoutQueue.enqueueTimeout(negotiationId, newTurnCount, AMBIENT_PARK_WINDOW_MS);
 
     this.logger.info('[NegotiationClaimTimeoutJob] AI agent countered, armed timeout for next speaker', {
