+++ backend/src/services/negotiation-polling.service.ts
@@ -5,7 +5,7 @@ import * as convSchema from '../schemas/conversation.schema';
 import * as dbSchema from '../schemas/database.schema';
 import { conversationDatabaseAdapter, ChatDatabaseAdapter } from '../adapters/database.adapter';
 import { negotiationTimeoutQueue } from '../queues/negotiation-timeout.queue';
-import { negotiationClaimTimeoutQueue } from '../queues/negotiation-claim-timeout.queue';
+import { negotiationClaimTimeoutQueue } from '../queues/negotiations/claim-timeout.queue';
 import { log } from '../lib/log';
 import type { NegotiationTurn, UserNegotiationContext, SeedAssessment } from '@indexnetwork/protocol';
 import { AMBIENT_PARK_WINDOW_MS } from '@indexnetwork/protocol';
